requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.1
django-bootstrap4==22.2
django-stdimage==6.0.1
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
whitenoise==6.2.0
